consumer/requirements.txt
pynsq==0.8.3
requests==2.25.0